services/extractor/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber
pymupdf==1.23.8

# AI/ML Libraries (Optional - will gracefully degrade if not available)
# openai==1.3.7
# azure-ai-formrecognizer==3.3.0
# azure-ai-textanalytics==5.3.0
# azure-storage-blob==12.19.0
# transformers==4.36.0
# torch>=2.2.0
# spacy==3.7.2
# scikit-learn==1.3.2

# NLP & Text Processing
nltk==3.8.1
# textract
python-dateutil==2.8.2
regex==2023.10.3

# Data Processing
pandas==2.1.4
numpy
openpyxl==3.1.2
xlrd==2.0.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
Pillow==10.1.0

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
